backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
faster-whisper==1.0.3
ffmpeg-python==0.2.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
# Опционально для speaker diarization (раскомментируйте при необходимости):
# whisperx==3.1.1
# pyannote.audio==3.1.1
# torch-audio==2.1.0

